+++ protocols/substreams/ethereum-rocketpool/abi/RocketDepositPool.json
