+++ protocols/substreams/ethereum-rocketpool/abi/RocketNetworkBalances.json
