+++ protocols/substreams/ethereum-rocketpool/src/abi/mod.rs
@@ -1,6 +1,6 @@
 #![allow(clippy::all)]
 
 pub mod rocket_dao_protocol_proposal;
-pub mod rocket_deposit_pool_v4;
-pub mod rocket_network_balances_v4;
+pub mod rocket_deposit_pool;
+pub mod rocket_network_balances;
 pub mod rocket_token_reth;
