+++ protocols/substreams/ethereum-rocketpool/src/abi/rocket_deposit_pool.rs
