+++ protocols/substreams/ethereum-rocketpool/src/abi/rocket_network_balances.rs
