+++ protocols/substreams/ethereum-rocketpool/substreams.yaml
@@ -20,7 +20,7 @@ binaries:
 network: mainnet
 
 params:
-  map_protocol_components: &starting_block 24479994 # Saturn I upgrade block (all v4 settings finalized)
+  map_protocol_components: &starting_block 24479994 # v1.4 (Saturn 1) upgrade block
   # Rocket Pool state values at the starting block. Verified against on-chain state at block 24479994.
   # To regenerate for a different starting block, run:
   #   RPC_URL=<your-archive-rpc> ./scripts/compute_initial_state.sh <block_number>
